restructure files into src folder
