+++ package.json
@@ -2,8 +2,17 @@
   "name": "jimp",
   "version": "0.2.28",
   "description": "An image processing library written entirely in JavaScript (i.e. zero external or native dependencies).",
-  "main": "jimp.js",
+  "main": "./src/jimp.js",
   "types": "./jimp.d.ts",
+  "tonicExampleFilename": "example.js",
+  "browser": "./browser/lib/jimp.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/oliver-moran/jimp.git"
+  },
+  "bugs": {
+    "url": "https://github.com/oliver-moran/jimp/issues"
+  },
   "scripts": {
     "lint": "npm run -s lint:src && npm run -s lint:tests && npm run -s lint:extra && echo Lint All Success",
     "lint:src": "xo *.js src/*.js",
@@ -81,15 +90,6 @@
     "uglify-js": "^3.4.5",
     "xo": "^0.21.1"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/oliver-moran/jimp.git"
-  },
-  "bugs": {
-    "url": "https://github.com/oliver-moran/jimp/issues"
-  },
-  "tonicExampleFilename": "example.js",
-  "browser": "browser/lib/jimp.js",
   "husky": {
     "hooks": {
       "pre-commit": "lint-staged"
