+++ src/jimp.js
