+++ src/modules/phash.js
