+++ src/modules/resize.js
