+++ src/modules/resize2.js
