+++ src/server.js
@@ -5,7 +5,7 @@ const express = require('express');
 
 const app = express();
 
-app.use(express.static(path.resolve(__dirname, './browser')));
+app.use(express.static(path.resolve(__dirname, '../browser')));
 app.listen(8080);
 
 console.log('Serving on http://127.0.0.1:8080');
